Item8307: Add caching to DirectedGraphWebMapPlugin The plugin is still slow the first time it renders a graph, but it is several orders of magnitude faster for subsequent views.
I also split off the code that finds the links between topics.
The undocumented "mapper" parameter lets you choose a different mapping
algorithm. The default algorithm is FindRenderedLinks. 
This is the same as it always was.

I implemented another mapper: Search. 
It is about three times slower than FindRenderedLinks on my test machine
which uses a vanilla config i.e. CGI, forking search and real RCS.
My main reason for checking it in is to help me test.
However, the Search mapper would be faster if a single search takes less time that the renderer takes to render one topic.



git-svn-id: http://svn.foswiki.org/trunk/DirectedGraphWebMapPlugin@9591 0b4bb1d4-4e5a-0410-9cc4-b2b747904278